refactor(screens): refactor show more screen add user section component
