--- src/components/showMore/Section/styled.ts
@@ -1,14 +1,0 @@
-import styled from '@emotion/native';
-
-export const SectionContainer = styled.View`
-  width: 100%;
-  flex-direction: column;
-  row-gap: 40px;
-  padding: 0 10px;
-  margin-top: 20px;
-`;
-
-export const Section = styled.View`
-  flex-direction: column;
-  row-gap: 20px;
-`;
