+++ src/components/showMore/ShowMoreCard/index.tsx
@@ -6,16 +6,13 @@ import { useTheme } from '@emotion/react';
 
 import { useNavigate, usePressingAnimation } from 'src/hooks';
 import { Icon, Text } from 'src/components';
+import { ShowMoreSectionItem } from 'src/constants/showMore';
 
 import * as S from './styled';
 
-export interface SectionItemProps {
-  name: string;
-  icon: string;
-  navigateUrl: string;
-}
+export interface SectionItemProps extends ShowMoreSectionItem {}
 
-export const SectionItem: React.FC<SectionItemProps> = ({ name, icon, navigateUrl }) => {
+export const ShowMoreCard: React.FC<SectionItemProps> = ({ name, icon, navigateUrl }) => {
   const theme = useTheme();
 
   const navigate = useNavigate();
@@ -30,17 +27,17 @@ export const SectionItem: React.FC<SectionItemProps> = ({ name, icon, navigateUr
       activeOpacity={1}
       onPress={() => navigate(navigateUrl)}
     >
-      <S.SectionItemWrapper style={[bgColorAnimatedStyle]}>
-        <S.SectionItem style={[scaleAnimatedStyle]}>
-          <S.SectionIconContainer>
+      <S.ShowMoreCardWrapper style={[bgColorAnimatedStyle]}>
+        <S.ShowMoreCardContainer style={[scaleAnimatedStyle]}>
+          <S.ShowMoreCardIconContainer>
             <Icon icon={icon} />
             <Text key={name} size={16}>
               {name}
             </Text>
-          </S.SectionIconContainer>
+          </S.ShowMoreCardIconContainer>
           <MaterialIcons name="chevron-right" size={30} color={theme.placeholder} />
-        </S.SectionItem>
-      </S.SectionItemWrapper>
+        </S.ShowMoreCardContainer>
+      </S.ShowMoreCardWrapper>
     </TouchableOpacity>
   );
 };
