+++ src/components/showMore/ShowMoreCard/styled.ts
@@ -2,19 +2,19 @@ import { Animated } from 'react-native';
 
 import styled from '@emotion/native';
 
-export const SectionItemWrapper = styled(Animated.View)`
+export const ShowMoreCardWrapper = styled(Animated.View)`
   border-radius: 16px;
   padding: 4px;
   padding-left: 0;
 `;
 
-export const SectionItem = styled(Animated.View)`
+export const ShowMoreCardContainer = styled(Animated.View)`
   flex-direction: row;
   align-items: center;
   justify-content: space-between;
 `;
 
-export const SectionIconContainer = styled.View`
+export const ShowMoreCardIconContainer = styled.View`
   flex-direction: row;
   align-items: center;
   justify-content: center;
