+++ src/components/showMore/ShowMoreSection/styled.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/native';
+
+export const ShowMoreSectionContainer = styled.View`
+  flex-direction: column;
+  row-gap: 20px;
+`;
