+++ src/components/showMore/index.ts
@@ -1,2 +1,4 @@
-export * from './Section';
-export * from './SectionItem';
+export * from './ShowMoreSection';
+export * from './ShowMoreCard';
+export * from './ShowMoreHeader';
+export * from './UserSection';
